Mostly minor bug fixes and improvements that have been exposed with the new TASKS.PSP search capability.
Fixed ParseException to not print it's Exception to stdout, since
this is needlessly verbose.

Fixed the TypeHelper to correctly identify subclasses of Collection
as valid arguments to the "all"/"exists"/"empty" operators.  Interally
added a new List type to make List-based operations more efficient than
using a Collection's Iterator.

Fixed the "this" operator to do better type checking, where previously
it would complain about an internal "PARSER ERROR".

Fixed "or" and "and" to trim out "true" and "false" arguments, which
allows simplification of statements such as (and x true) to just (x).
Also fixed the "apply" operator and MethodOp to similarily identify
the "true" and "false" operator arguments and simplify the expressions,
such as (apply x true) is just (true).

Fixed the paren-based "toString()" to correctly quote constants such
as (const "int" "5").